Update VforMSF.py
Added a check for certain filetypes when using the "generate" module so that an acceptable format is loaded. Specifically there was trouble with .php, .apk, and .jsp file types needing to be a different format such as "raw". 

I may implement additional features for more advanced payload generation in the future.